update overview data

+++ service/multer.js
@@ -101,9 +101,12 @@ const getResult = async (req, res, next) => {
 };
 
 function deleteImage(imageName){
-    fs.unlink(path.resolve('public/images/'+imageName),(err)=>{
-        if(err) err_service.errorNotification(err,'delete image')
-    })
+    if(fs.existsSync(path.resolve('public/images/'+imageName))){
+        fs.unlink(path.resolve('public/images/'+imageName),(err)=>{
+            if(err) err_service.errorNotification(err,'delete image')
+        })
+    }
+
 }
 
   module.exports = {
